[Connectors] Migrate notion types to @dust-tt/types
Requirement to properly handle types for https://github.com/dust-tt/tasks/issues/540
