+++ connectors/src/connectors/notion/temporal/activities.ts
@@ -3,6 +3,7 @@ import type {
   ModelId,
   NotionGarbageCollectionMode,
 } from "@dust-tt/types";
+import type { PageObjectProperties, ParsedNotionBlock } from "@dust-tt/types";
 import { assertNever, makeStructuredDataTableName } from "@dust-tt/types";
 import { isFullBlock, isFullPage, isNotionClientError } from "@notionhq/client";
 import type { PageObjectResponse } from "@notionhq/client/build/src/api-endpoints";
@@ -39,10 +40,6 @@ import {
   updateAllParentsFields,
 } from "@connectors/connectors/notion/lib/parents";
 import { getTagsForPage } from "@connectors/connectors/notion/lib/tags";
-import type {
-  PageObjectProperties,
-  ParsedNotionBlock,
-} from "@connectors/connectors/notion/lib/types";
 import { dataSourceConfigFromConnector } from "@connectors/lib/api/data_source_config";
 import { concurrentExecutor } from "@connectors/lib/async_utils";
 import {
